% Generated by roxygen2 (4.0.2): do not edit by hand
\name{deleteMetadata}
\alias{deleteMetadata}
\title{Delete Model Metadata from the InfoSphere Information Governance Catalog for the given App}
\usage{
deleteMetadata(name = "R")
}
\arguments{
\item{name}{The name of the app}
}
\description{
Delete Model Metadata from the InfoSphere Information Governance Catalog for the given App
}

