% Generated by roxygen2 (4.0.2): do not edit by hand
\name{exportMetadata}
\alias{exportMetadata}
\title{Export Model Metadata to the InfoSphere Information Governance Catalog as PMML}
\usage{
exportMetadata(model, name, data)
}
\arguments{
\item{model}{The predictive model}

\item{name}{The name of the model}

\item{data}{The Dataframe for the model}
}
\description{
Export Model Metadata to the InfoSphere Information Governance Catalog as PMML
}

